class Counter separetedd to own file
